--- lib/AuthTrial.dart
@@ -1,17 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-
-class AuthService {
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-//  User _userFromFirebaseUser()
-
-  Future signInAnon() async {
-    try{
-      AuthResult result = await _auth.signInAnonymously();
-      FirebaseUser user = result.user;
-      return user;
-  }catch (e){
-      print(e.toString());
-      return null;
-  }
-}
-}
