+++ lib/Game/Garden2.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:lwk/Garden.dart';
-import 'package:lwk/Kitchen.dart';
+import 'package:lwk/Game/Garden3.dart';
+import 'package:lwk/Game/Storage.dart';
 import 'package:lwk/main.dart';
-class Garden3 extends StatelessWidget {
+class Garden2 extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -14,12 +14,12 @@ class Garden3 extends StatelessWidget {
             SizedBox(height: 30),
             Text('Garden',style: TextStyle(fontSize: 50,decoration: TextDecoration.underline),),
             Container(
-              child: Text('Great! You Earned'),
+              child: Text('Harvested!'),
               padding: EdgeInsets.all(10),
               decoration: BoxDecoration(
-                color: Colors.green[300],
-                border: Border.all(width: 2.0),
-                borderRadius: BorderRadius.all(Radius.circular(20.0)),
+                  color: Colors.green[300],
+                  border: Border.all(width: 2.0),
+                  borderRadius: BorderRadius.all(Radius.circular(20.0)),
               ),
             ),
             SizedBox(height: 100),
@@ -27,15 +27,15 @@ class Garden3 extends StatelessWidget {
             SizedBox(height: 40),
             RaisedButton(
               onPressed: (){
-                Navigator.pushReplacement(context, ScaleRoute(page: Kitchen()));
+                Navigator.pushReplacement(context, ScaleRoute(page: Storage()));
               },
-              child: Text('Go back home'),
+              child: Text('Store it'),
             ),
             RaisedButton(
               onPressed: (){
-                Navigator.pushReplacement(context, ScaleRoute(page: Garden()));
+                Navigator.pushReplacement(context, ScaleRoute(page: Garden3()));
               },
-              child: Text('Go plant more'),
+              child: Text('Sell it'),
             ),
           ],
         ),
