+++ lib/main.dart
@@ -1,103 +1,36 @@
 import 'package:flutter/material.dart';
-import 'package:lwk/AuthLogin.dart';
-import 'package:lwk/Kitchen.dart';
-import 'package:lwk/Storage.dart';
-import 'package:lwk/Telephone.dart';
-import 'package:lwk/Garden.dart';
-import 'package:lwk/Bedroom.dart';
+import 'package:lwk/Auth_service.dart';
+import 'package:lwk/Game/Start.dart';
+import 'package:lwk/Wrapper.dart';
+import 'package:provider/provider.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
-void main() => runApp(MaterialApp(
-  theme: ThemeData(
-      primaryTextTheme: TextTheme(title: TextStyle(fontSize: 40)),
-      //dialogTheme: DialogTheme(contentTextStyle: TextStyle(fontSize: 20)),
-      fontFamily: 'tStyle',
-      textTheme: TextTheme(
-        title: TextStyle(fontSize: 40),
-        subhead: TextStyle(fontSize: 30),
-        body1: TextStyle(fontSize: 35),
-        button: TextStyle(fontSize: 30),
-      ),
-      buttonTheme: ButtonThemeData(buttonColor: Colors.pink[300]),
-    ),
-//  home: Start(),
-  initialRoute: '/',
-  routes: {
-    '/': (context) => Start(),
-  },
-)
-);
 
-class Start extends StatefulWidget {
-  @override
-  _StartState createState() => _StartState();
-}
+void main() => runApp(Main());
 
-class _StartState extends State<Start> {
+class Main extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: Text('The COVID-19 Survival Guide')),
-      body: Stack(
-        children: <Widget>[
-//          Container(
-//            decoration: BoxDecoration(image: DecorationImage(image: Image'imageBckg.jpg',fit: BoxFit.cover)),
-//          ),
-          Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: <Widget>[
-                SizedBox(height: 30,),
-                Text('Welcome!',style: TextStyle(fontSize: 50,decoration: TextDecoration.underline),),
-                SizedBox(height: 100),
-                Text('Where do you wanna go?'),
-                SizedBox(height: 40),
-                RaisedButton(
-                  onPressed: (){
-                    Navigator.push(context, ScaleRoute(page: SignIn()));
-                    //Navigator.push(context, ScaleRoute(page: Kitchen()));
-                  },
-                  child: Text('Kitchen'),
-                ),
-                RaisedButton(
-                  onPressed: (){
-                    Navigator.push(context, ScaleRoute(page: Storage()));
-                  },
-                  child: Text('Storage'),
-                ),
-                RaisedButton(
-                  onPressed: (){
-                    Navigator.push(context, ScaleRoute(page: Telephone()));
-                  },
-                  child: Text('Telephone'),
-                ),
-                RaisedButton(
-                  onPressed: (){
-                    Navigator.push(context, ScaleRoute(page: Garden()));
-                  },
-                  child: Text('Garden'),
-                ),
-                RaisedButton(
-                  onPressed: () async {
-                    await showDialog<void>(
-                        context: context,
-                        builder: (BuildContext context) {
-                          return AlertDialog(
-                            content: Text('I\'m glad you are well-rested.'),
-                            actions: <Widget>[FlatButton(
-                              onPressed: (){Navigator.of(context).pop();},
-                              child: Text('Ok'),
-                            )],
-                          );
-                        }
-                    );
-                    Navigator.push(context, ScaleRoute(page: Bedroom()));
-                  },
-                  child: Text('Sleep'),
-                ),
-              ],
-            ),
+    return  StreamProvider<FirebaseUser>.value(
+      value: AuthService().user,
+      child: MaterialApp(
+        theme: ThemeData(
+          primaryTextTheme: TextTheme(title: TextStyle(fontSize: 40)),
+          //dialogTheme: DialogTheme(contentTextStyle: TextStyle(fontSize: 20)),
+          fontFamily: 'tStyle',
+          textTheme: TextTheme(
+            title: TextStyle(fontSize: 40),
+            subhead: TextStyle(fontSize: 30),
+            body1: TextStyle(fontSize: 35),
+            button: TextStyle(fontSize: 30),
           ),
-        ]
+          buttonTheme: ButtonThemeData(buttonColor: Colors.pink[300]),
+        ),
+        //home: Wrapper(),
+        routes: {
+          '/': (context) => Wrapper(),
+          'start': (context) => Start(),
+        },
       ),
     );
   }
