+++ pubspec.yaml
@@ -27,6 +27,7 @@ dependencies:
 dev_dependencies:
   firebase_auth: ^0.16.0
   cloud_firestore: ^0.13.5
+  provider: ^3.1.0
   flutter_test:
     sdk: flutter
